+++ variables.js
@@ -151,5 +151,6 @@ botellademiel: "0",
 twsubs: "0",
 ElixirOscuro: "0",
 cascominero: "0",
+himatospell: "0",
 };
 module.exports = variables;